+++ packages/boot/src/components/root/single-page.tsx
@@ -55,6 +55,7 @@ export default function RootSinglePage() {
 								[ styles[ 'has-canvas' ] ]:
 									!! canvas || canvas === null,
 								[ styles[ 'has-full-canvas' ] ]: isFullScreen,
+								'boot-is-full-canvas': isFullScreen,
 							}
 						) }
 					>